@@ -1,13 +1,10 @@
-import Vue from 'vue'
-import './plugins/sara'
-import './plugins/axios'
+import { createApp } from 'vue'
 import App from './App.vue'
-import './assets/tailwind.css'
-import router from './router'
 
-Vue.config.productionTip = false
+import router from "./router";
 
-new Vue({
-  router,
-  render: h => h(App)
-}).$mount('#app')
+import './style.css'
+
+createApp(App).
+    use(router).
+    mount('#app')
