@@ -1,11 +1,12 @@
 <template>
-  <div>
+  <div class="relative bg-white">
+    <app-header/>
     <router-view/>
+    <app-footer/>
   </div>
 </template>
 
-<script>
-export default {
-  name: 'App'
-}
+<script setup>
+import AppHeader from "./components/AppHeader.vue"
+import AppFooter from "./components/AppFooter.vue"
 </script>
