@@ -1,128 +1,38 @@
 <template>
   <div>
-    <div
-      class="
-        mt-10
-        mx-auto
-        py-10
-        max-w-7xl
-        px-4
-        sm:mt-12 sm:px-6
-        md:mt-16
-        lg:mt-20 lg:px-8
-        xl:mt-28
-      "
-    >
+    <div class="mt-10 mx-auto py-10 max-w-7xl px-4 sm:mt-12 sm:px-6 md:mt-16 lg:mt-20 lg:px-8 xl:mt-28">
       <div class="sm:text-center lg:text-left">
         <h1 class="text-4xl tracking-tight font-extrabold text-gray-900">
-          <span class="block xl:inline">想要</span>
-          <span class="block text-violet-600 xl:inline">分享文字片段</span>
-          <span class="block xl:inline">卻找不到地方貼貼剪剪嗎？</span>
+          <span class="block xl:inline">這裡是</span>
+          <span class="block text-indigo-600 xl:inline">前端 Web 模板</span>
         </h1>
         <p class="mt-3 text-base text-gray-500">
-          我們提供一個簡單好用的地方，讓你可以在這裡打打字，並分享給朋友
+          Vite + Vue.js 3 + Tailwind CSS，為了快速發展專案而生，請自行修改內容。
         </p>
         <div class="mt-5 sm:mt-8 sm:flex sm:justify-center lg:justify-start">
           <div class="mt-3 sm:mt-0 sm:ml-3">
-            <button
-              class="
-                w-full
-                flex
-                items-center
-                justify-center
-                px-8
-                py-3
-                border border-transparent
-                text-base
-                font-medium
-                rounded-md
-                text-violet-700
-                bg-violet-100
-                hover:bg-violet-200
-                md:py-4 md:text-lg md:px-10
-              "
-              @click="goWriter"
-            >
-              開始打字
-            </button>
+            <a href="https://github.com/web-tech-tw/template.inte" class="w-full flex items-center justify-center px-8 py-3 border border-transparent text-base font-medium rounded-md text-indigo-700 bg-indigo-100 hover:bg-indigo-200 md:py-4 md:text-lg md:px-10">
+              template.inte GitHub
+            </a>
           </div>
         </div>
       </div>
     </div>
-    <div
-      class="mt-10 h-14 bg-gradient-to-r from-violet-500 to-fuchsia-500"
-    ></div>
-    <div
-      class="
-        mt-10
-        mx-auto
-        py-10
-        max-w-7xl
-        px-4
-        sm:mt-12 sm:px-6
-        md:mt-16
-        lg:mt-20 lg:px-8
-        xl:mt-28
-      "
-    >
+    <div class="mt-10 mx-auto py-10 max-w-7xl px-4 sm:mt-12 sm:px-6 md:mt-16 lg:mt-20 lg:px-8 xl:mt-28">
       <div class="sm:text-center lg:text-right">
         <h1 class="text-4xl tracking-tight font-extrabold text-gray-900">
-          <span class="block xl:inline">你</span>
-          <span class="block text-violet-600 xl:inline">已經拿到</span>
-          <span class="block xl:inline">片段代碼了嗎？</span>
+          <span class="block xl:inline">你是</span>
+          <span class="block text-indigo-600 xl:inline">後端</span>
+          <span class="block xl:inline">工程師嗎？</span>
         </h1>
         <p class="mt-3 text-base text-gray-500">
-          這裡這裡，往這裡輸入你拿到的代碼，我來幫你換成片段資料！
+          你走錯地方了，下方連結才是後端 API 模板。
         </p>
         <div class="mt-5 sm:mt-8 sm:flex sm:justify-center lg:justify-end">
-          <div class="rounded-md shadow mr-1">
-            <input
-              class="
-                w-full
-                flex
-                items-center
-                justify-center
-                px-8
-                py-3
-                border border-transparent
-                text-base
-                font-medium
-                rounded-md
-                text-black
-                bg-white
-                hover:bg-gray-100
-                my-2
-                sm:my-0
-                md:py-4 md:text-lg md:px-10
-              "
-              type="text"
-              placeholder="0code-xxxx"
-              v-model="code"
-              @keydown.enter="goReader"
-            />
-          </div>
           <div class="rounded-md shadow">
-            <button
-              class="
-                w-full
-                flex
-                items-center
-                justify-center
-                px-8
-                py-3
-                border border-transparent
-                text-base
-                font-medium
-                rounded-md
-                text-white
-                bg-violet-600
-                hover:bg-violet-700
-                md:py-4 md:text-lg md:px-10
-              "
-              @click="goReader"
-            >
-              查詢
-            </button>
+            <a href="https://github.com/web-tech-tw/template.recv" class="w-full flex items-center justify-center px-8 py-3 border border-transparent text-base font-medium rounded-md text-white bg-indigo-600 hover:bg-indigo-700 md:py-4 md:text-lg md:px-10">
+              template.recv GitHub
+            </a>
           </div>
         </div>
       </div>
@@ -132,20 +42,6 @@
 
 <script>
 export default {
-  name: "HomeView",
-  data: () => ({
-    code: "",
-  }),
-  methods: {
-    goWriter() {
-      this.$router.push("/writer");
-    },
-    goReader() {
-      if (!this.code) {
-        return;
-      }
-      this.$router.push(`/r/${this.code}`);
-    },
-  },
-};
+  name: 'HomeView',
+}
 </script>
